requirements.txt
pytesseract
opencv-python
numpy
mss
Pillow
pygetwindow